[Rebase-Check] Code dump of work done for settings migration tool
Had this in my stash pile for a while, should probably commit so I don't lose
it. Not at all usable, still contains snippets of config files used for
reference.

Rebase-Check is a reminder to edit/drop this commit as appropriate when
rebasing for merging into master, based on what has been implemented.